Fix two real publish-pipeline bugs found running it for the first time
- publish_digital: Etsy caps file names at 70 chars; a long design_id
  plus a size suffix (e.g. "...sage-minimal_letter_8.5x11") could
  exceed that and 400 on upload. Truncate the design_id half.
- printify_client: create_product's default angle=0.0 (float) fails
  Printify's validation, which requires an integer. Default to 0 (int).

Also adds EtsyClient.delete_listing, used to clean up an orphaned
draft left by the file-name bug before the fix landed.

+++ integrations/printify_client.py
@@ -106,7 +106,7 @@ class PrintifyClient:
         scale: float = 1.0,
         x: float = 0.5,
         y: float = 0.5,
-        angle: float = 0.0,
+        angle: int = 0,
     ) -> dict:
         body = {
             "title": title,
